Fix missing dependencies and refactor so that it is easier to have a test that runs almost all imports
