+++ package.json
@@ -28,6 +28,7 @@
     "access": "public"
   },
   "dependencies": {
+    "@octokit/app": "^3.0.0",
     "@octokit/rest": "^16.3.2",
     "async": "^2.6.1",
     "aws-sdk": "^2.395.0",
